Searching Controller finished and ui suggestions real time implemented in ui
