+++ app/src/main/res/layout/activity_main.xml
@@ -28,7 +28,6 @@
             android:layout_marginEnd="10dp"
             android:layout_marginBottom="16dp"
             android:background="#FFFFFF"
-            tools:itemCount="5"
             android:text="@string/app_name"
             android:contentDescription="@string/app_name">
